[initfs] Change hw-ramdisk into a set of tools for generating initrds
The old way of extracting and adding manually files to initrd based on
mer-ramdisk does not scale well when we need to add binaries with library
dependencies to any given initrd.

Replaced the old ready made ramdisk cpio archive with set of initrd
creation tools based on moslo project. These tools can be used to
generate the initrd in any downstream project by providing a list of
files that are needed. The tools will automatically add right
versions of needed libraries to the initrd to make sure your binary
tools always work.

The template contains already busybox-static with symlinks added, so
in simplest case you only provide your init script to the tools
and combine the final rootfs.cpio.gz with your kernel zImage.

The sequence to create a initrd contains now 2 steps:
1. initialize-ramdisk.sh
   * Initialize needed directories and file lists for moslo-build
2. moslo-build.sh
   * Build the initrd based on files in initfs/skeleton/ and given params

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>