[moslo] Make kernel modules a voluntary option
Currently moslo-build.sh requires "-m" option to be given for
included kernel modules. This can be quite cumbersome during
kernel development as you in most cases are happy with a kernel
without modules in initrd. Also with kernels built with HADK
environment cannot provide the modules during initrd creation
time, but have to add them later if needed.

So fixed this by making the kernel module addition depend on if "-m"
option was provided.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>