[moslo] Recursive library dependency search with objdump. Fixes JB#28364.
The objdump based search was not working correctly as ld-linux.so.2 on
i486 scratchbox2 targets doesn't really find correct libraries. This means
that only first level of dependencies were found but further levels were
missed in some cases.

To fix this, a recursive completely objdump based search for non-runnable
binaries was implemented.

This is slightly slower than ld-linux, but seems to work robustly
on i486 scratchbox2.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>