[moslo] Don't duplicate libraries. Fixes JB#29318
Currently also the actual library is getting included in addition
to the symlink. However moslo-build.sh only copies the SONAME symlink
as actual binary copy of the library. Therefore copying the actual
library is redundant. Replaced the library path checking with a
ldconfig based path search that only includes the symlink path.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>